+++ src/component/Subtotal.js
@@ -1,8 +1,8 @@
 import React from "react";
-import "./Subtotal.css";
+import "../assets/styles/Subtotal.css";
 import CurrencyFormat from "react-currency-format";
-import { useStateValue } from "./StateProvider";
-import { getBasketTotal } from "./reducer";
+import { useStateValue } from "../StateProvider";
+import { getBasketTotal } from "../reducer";
 
 const Subtotal = () => {
   const [{ basket }, dispatch] = useStateValue();
